# Histograms: optimize floatBucketIterator for common case