This change improves performance of `floatBucketIterator` by introducing a special case branch to `floatBucketIterator.Next` which simplifies handling of the common case (`i.schema == i.targetSchema`) by avoiding all the logic of merging buckets into the target schema.

Benchmark results: 
```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/histogram
                                   │ before.txt  │              after.txt              │
                                   │   sec/op    │   sec/op     vs base                │
FloatHistogramAllBucketIterator-10   41.29µ ± 1%   40.46µ ± 0%   -2.02% (p=0.000 n=10)
FloatHistogramDetectReset-10         21.96µ ± 1%   17.28µ ± 0%  -21.30% (p=0.000 n=10)
geomean                              30.11µ        26.44µ       -12.18%
```
Note that the improvement in `FloatHistogramAllBucketIterator` is much smaller, apparently because it is dominated by `getBound` calls.